Add files via upload
Hi,  Cohen,
This is the updated homework of 4,5,6 and 7. I was trying to combine the three method to one in MergeSort class of homework 6. But it didnt work,  I really cant figure it out. The code looks right and it works before. I didnt change anything.  Sorry about send this so late and thank you.